requirements.txt
discord.py==2.4.0
sqlalchemy==2.0.0
cryptography==43.0.0
python-dotenv==1.0.0
asyncio==3.4.3